requirements.txt
deffcode
twelvelabs
python-dotenv
